Move more tests to their own .rs file
